Player Skill(3) - AOE - 힐링 스킬 구현 - Skill replicate error 해결 - 코드 정리 필요

+++ FinalProject/Source/FinalProject/Character/WizardPlayerState.h
@@ -21,34 +21,34 @@ public:
 	EWizardJob mJob = EWizardJob::Blue;
 
 	UPROPERTY(EditAnywhere)
-	int32 mNormalAttackPoint = 20.f;
+	float mNormalAttackPoint = 20.f;
 	UPROPERTY(EditAnywhere)
 	float mNormalAttackDistance = 350.f;
 
 	UPROPERTY(EditAnywhere)
-	int32 mFirstSkillAttackDamage = 100.f;
+	float mFirstSkillAttackDamage = 100.f;
 	UPROPERTY(EditAnywhere)
 	float mFirstSkillAttackDistance = 500.f;
 	UPROPERTY(EditAnywhere)
-	int32 mSecondSkillAttackDamage = 100.f;
+	float mSecondSkillAttackDamage = 100.f;
 	UPROPERTY(EditAnywhere)
 	float mSecondSkillAttackDistance = 500.f;
 	UPROPERTY(EditAnywhere)
-	int32 mThirdSkillAttackDamage = 100.f;
+	float mThirdSkillAttackDamage = 100.f;
 	UPROPERTY(EditAnywhere)
 	float mThirdSkillAttackDistance = 500.f;
 	UPROPERTY(EditAnywhere)
-	int32 mFourthSkillAttackDamage = 100.f;
+	float mFourthSkillAttackDamage = 100.f;
 	UPROPERTY(EditAnywhere)
 	float mFourthSkillAttackDistance = 500.f;
 
 	UPROPERTY(EditAnywhere)
-	int32 mArmorPoint = 10.f;
+	float mArmorPoint = 10.f;
 
 	UPROPERTY(EditAnywhere)
-	int32 mHPMax = 100.f;
+	int32 mHPMax = 100;
 	UPROPERTY(EditAnywhere)
-	int32 mMPMax = 100.f;
+	int32 mMPMax = 100;
 };
 
 
@@ -70,28 +70,28 @@ public:
 	EWizardJob mJob;
 
 
-	int32 mNormalAttackPoint;
+	float mNormalAttackPoint;
 	float mNormalAttackDistance;
 
-	int32 mFirstSkillAttackDamage;
+	float mFirstSkillAttackDamage;
 	float mFirstSkillAttackDistance;
-	int32 mSecondSkillAttackDamage;
+	float mSecondSkillAttackDamage;
 	float mSecondSkillAttackDistance;
-	int32 mThirdSkillAttackDamage;
+	float mThirdSkillAttackDamage;
 	float mThirdSkillAttackDistance;
-	int32 mFourthSkillAttackDamage;
+	float mFourthSkillAttackDamage;
 	float mFourthSkillAttackDistance;
 
 
-	int32 mArmorPoint;
+	float mArmorPoint;
 
 	
-	int32 mHP;
+	float mHP;
 
 	int32 mHPMax;
 
 	
-	int32 mMP;
+	float mMP;
 
 	int32 mMPMax;
